new model request, new parser

+++ app/controllers/service.js
@@ -1,6 +1,8 @@
 const mongoose = require('mongoose');
 const Service = mongoose.model('Service');
 const Category = mongoose.model('Category');
+const Order = mongoose.model('Order');
+const parserError = require('../helpers/parserError');
 
 const {baseHref} = require('../../config/app');
 
@@ -10,6 +12,7 @@ const unlinkAsync = promisify(fs.unlink);
 
 const getAll = (req, res) => {
     Service.find()
+        .populate('category')
         .exec()
         .then(service => res.json(service))
         .catch(err => res.status(500).json(err));
@@ -28,7 +31,7 @@ const create = (req, res) => {
             return Category.findOneAndUpdate({_id: service.category}, {$push: {service: service._id}});
         })
         .then(category => {
-            res.json(category);
+            res.json(parserError(category));
         })
         .catch(err => res.status(500).json(err));
 
@@ -53,19 +56,29 @@ const update = (req, res) => {
             return Category.findOneAndUpdate({_id: newService.category}, {$push: {service: newService.id}});
         })
         .then(category => {
-            res.json(category);
+            res.json(parserError(category));
         })
         .catch(err => res.status(500).json(err));
 };
 
 const remove = (req, res) => {
-    Service.findOneAndRemove({_id: req.params.id})
+    Service.findOne({_id: req.params.id})
         .exec()
+        .then(service => {
+            return Order.find({service: service._id});
+        })
+        .then(order => {
+            if (order.length > 0) {
+                res.status(500).json('Запрет на удаление! Имеются привязанные записи.')
+            } else {
+                return Service.findOneAndRemove({_id: req.params.id});
+            }
+        })
         .then(service => {
             deleteImage(service);
             return Category.findOneAndUpdate({_id: service.category}, {$pull: {service: service._id}})
         })
-        .then(category => res.json(category))
+        .then(category => res.json(parserError(category)))
         .catch(err => res.status(500).json(err));
 };
